took away the size
t

+++ src/regrf.c
@@ -19,9 +19,9 @@
 void simpleLinReg(int nsample, double *x, double *y, double *coef,
 		  double *mse, int *hasPred);
 
-void ran_multinomial (size_t K, int N, 
+void ran_multinomial (int K, int N, 
                       double p[], int *coeffs){
-  size_t k;
+  int k;
   double norm  = 0.0;
   double sum_p = 0.0;
